code update

+++ lcci/05.06.Convert Integer/Solution.swift
@@ -1,5 +1,5 @@
 class Solution {
     func convertInteger(_ A: Int, _ B: Int) -> Int {
-        return (A ^ B).nonzeroBitCount
+        return (Int32(A) ^ Int32(B)).nonzeroBitCount
     }
 }
